// apps/api/src/generated/prisma/enums.ts
export const UserStatus = {
  ACTIVE: 'ACTIVE',
  DISABLED: 'DISABLED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const ProjectRole = {
  OWNER: 'OWNER',
  MANAGER: 'MANAGER',
  MEMBER: 'MEMBER',
  VIEWER: 'VIEWER'
} as const

export type ProjectRole = (typeof ProjectRole)[keyof typeof ProjectRole]


export const ProjectKind = {
  GENERAL: 'GENERAL',
  COMPANION: 'COMPANION'
} as const

export type ProjectKind = (typeof ProjectKind)[keyof typeof ProjectKind]


export const TaskStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  CLAIMED: 'CLAIMED',
  IN_PROGRESS: 'IN_PROGRESS',
  SUBMITTED: 'SUBMITTED',
  REVISION_REQUESTED: 'REVISION_REQUESTED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  DISPUTED: 'DISPUTED',
  REMOVED: 'REMOVED'
} as const

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus]